Fix potential simultaneous `apply()` calls from `.get()` call(s)
`LinearizedCore._rebuild` can be called from two locations
(`LinearizedCore._update` & `LinearizedCore.get` upon a retry failing)
causing the `applyFunction` multiple times simultaneously. This will
throw the 'Can only append to a LinearizedCore inside of an apply
function' in `LinearizedCore.append` and the LinearizedCore's
properties, such as `this.status`, to potentially be partially updated.

The fix was to add a mutexify lock for `LinearizedCore._rebuild` to
block it from being run simultaneously. A test with an apply function
with a delay added to ensure testing applys running at the same time.