Add localInput to inputs automatically
Prior to this, one needed to add the local input twice; once as
`localInput` in the constructor and once in the `inputs` array option in
the constructor.

This commit automatically adds the local input to the autobase's inputs
creating feature parity with the local output including dynamically
adding/removing a local input.