Increase `ackInterval` in 'autoack - value beneath null values' test
Previously the test could fail when an ack triggered during checks for
`sync()`. The ack would complete before the final promise resolved in
`sync()` causing the length stored before the 1s timeout to be the final
resolved length that is supposed to be different.

This is just a timing issue where the 'before' is remembered after the
'after' is complete.

100ms didn't flake after ~195 runs mining the test.