Validate fork system length against prologue
If in `apply` the `host.system.core.key` is used as the system key to
fork from, then the system will fork once correctly, setting the
prologue for the new indexers, but then will fail on reapplying post
fork.

While this might be discouraged as a pattern, the prologue should always
be behind the system fork length regardless so should be validated.