trace apply (#232)
* store trace on the nodes

* Add `trace` field to oplog v2

* Add block `trace` property to linearizer block

* Request tip trace view blocks before processing update

All block indexes are requested asynchronously to warm up blocks
`get`ted in the `apply` function allowing it to run faster. Block
requests are not awaited and can throw.

* Warm up blocks for indexed blocks too

* Add test verifying MAX_TRACE_PER_VIEW is max blocks traced

* Use `core.download()` instead of `core.get()`s for warmup

Range is destroyed after applying as it is no longer needed.

* Refactor to download ranges once per view in warmup

* Skip then warning when trace view is OOB for current set of views

* Remove lingering `warmups` range destroy calls

* Add test that `trace` entry for non-existent index doesnt leave request

Keeping the replication stream alive, all peers are checked to see if
there is any inflight requests.

* Skip tracing non-signed blocks in views

The MAX_TRACE_PER_VIEW test was revised to avoid requesting unindexed
blocks since it was previously appending a large batch to exceed the max
trace per view amount.

* Add test to verify new views are skipped

The new `v2` view is skipped because the core's `signedLength` is `0`
until indexed.

* Unify start & end of trace across writable & optimistic applies

Instead of having two ways of starting the trace, a single trace is used
at the beginning of the for loop for apply batches.

* Only define `onseq` for atoms, not batch sessions

`onseq` would be `null` anyways for the batch session so better to just
not define it.

* Narrow requirement for tracing to only if the node's core is writable

An optimistic append should only be traced if the core for the node is
writable since only the peer appending the node should be tracing it.

* Define reserved trace views (sys & encryption) separate from user views

Enables future internal views for autobase to also be traced without
using reserved indexes in a `trace` array. This structure also mirrors
the checkpoint struct, just without a signature and length.

Added support for encryption view as part of this refactor.

---------

Co-authored-by: Sean Zellmer <sean@lejeunerenard.com>